Integration for player

See merge request compsci308_2019spring/voogasalad_nomergeconflicts!129

+++ src/Configurations/src/Configs/EnemyPackage/EnemyBehaviors/Stealthy.java
@@ -18,7 +18,7 @@ public class Stealthy extends EnemyBehavior{
     }
 
     @Override
-    public void update(double ms) {
+    public void update(long ms) {
         //TODO
     }
     @Override
